Avoid activating conda environment
Recently it seems that some conda scripts are, under certain
circumstances, not generated in a way so that their shebang points to
the absolute path of conda's own Python binary.

Since we just need to make a few calls to conda ingredients, we can just
call these by their path.